feat: improve train details
